requirements.txt
PyPDF2 
pdfplumber 
python-docx
openai
python-dotenv
fastapi
uvicorn
python-multipart